People were reporting two issues:
* when searching for hotel, if they went back to the search portal and search again - they would get an error
* when doing air search and going back to search again they would get a 'stuck' summary page and not be able to checkout

Digging into this Kamie noticed the code was throwing a 403 error in the Javascript console log. These were being thrown from the application.cfc if a user was not authenticated correctly. Debugging it seems the users cookies were correctly set for the first request, but if they went back and searched again, their cookies would be gone and they would not be authenticated. Dumping to the console - it seems like the cookies were being overwritten. They are named the same in both search and booking.

I have replaced the second cookie check for the cfcRequest with a session check instead. The session check is set in the first request
and should persist as long as the user doesn't close their browser.

I have tweaked my local code to mimic what would occur in production and no longer get an error when going back and forth.